+++ Source/ASCellNode.mm
@@ -352,6 +352,24 @@
   return NO;
 }
 
+- (BOOL)shouldUseUIKitCell
+{
+  return NO;
+}
+
+@end
+
+#pragma mark -
+#pragma mark ASWrapperCellNode
+
+// TODO: Consider if other calls, such as willDisplayCell, should be bridged to this class.
+@implementation ASWrapperCellNode : ASCellNode
+
+- (BOOL)shouldUseUIKitCell
+{
+  return YES;
+}
+
 @end