+++ Source/Details/ASDataController.mm
@@ -17,6 +17,8 @@
 
 #import <AsyncDisplayKit/ASDataController.h>
 
+#include <atomic>
+
 #import <AsyncDisplayKit/_ASHierarchyChangeSet.h>
 #import <AsyncDisplayKit/_ASScopeTimer.h>
 #import <AsyncDisplayKit/ASAssert.h>
@@ -54,6 +56,8 @@ NSString * const ASCollectionInvalidUpdateException = @"ASCollectionInvalidUpdat
 
 typedef dispatch_block_t ASDataControllerCompletionBlock;
 
+typedef void (^ASDataControllerSynchronizationBlock)();
+
 @interface ASDataController () {
   id<ASDataControllerLayoutDelegate> _layoutDelegate;
 
@@ -69,6 +73,11 @@ typedef dispatch_block_t ASDataControllerCompletionBlock;
   
   BOOL _initialReloadDataHasBeenCalled;
 
+  BOOL _synchronized;
+  NSMutableSet<ASDataControllerSynchronizationBlock> *_onDidFinishSynchronizingBlocks;
+  
+  std::atomic<int> _editingTransactionQueueCount;
+  
   struct {
     unsigned int supplementaryNodeKindsInSections:1;
     unsigned int supplementaryNodesOfKindInSection:1;
@@ -112,6 +121,9 @@ typedef dispatch_block_t ASDataControllerCompletionBlock;
   _nextSectionID = 0;
   
   _mainSerialQueue = [[ASMainSerialQueue alloc] init];
+
+  _synchronized = YES;
+  _onDidFinishSynchronizingBlocks = [NSMutableSet set];
   
   const char *queueName = [[NSString stringWithFormat:@"org.AsyncDisplayKit.ASDataController.editingTransactionQueue:%p", self] cStringUsingEncoding:NSASCIIStringEncoding];
   _editingTransactionQueue = dispatch_queue_create(queueName, DISPATCH_QUEUE_SERIAL);
@@ -440,14 +452,7 @@ typedef dispatch_block_t ASDataControllerCompletionBlock;
 - (BOOL)isProcessingUpdates
 {
   ASDisplayNodeAssertMainThread();
-  if (_mainSerialQueue.numberOfScheduledBlocks > 0) {
-    return YES;
-  } else if (dispatch_group_wait(_editingTransactionGroup, DISPATCH_TIME_NOW) != 0) {
-    // After waiting for zero duration, a nonzero value is returned if blocks are still running.
-    return YES;
-  }
-  // Both the _mainSerialQueue and _editingTransactionQueue are drained; we are fully quiesced.
-  return NO;
+  return _mainSerialQueue.numberOfScheduledBlocks > 0 || _editingTransactionQueueCount > 0;
 }
 
 - (void)onDidFinishProcessingUpdates:(nullable void (^)())completion
@@ -466,9 +471,37 @@ typedef dispatch_block_t ASDataControllerCompletionBlock;
   }
 }
 
+- (BOOL)isSynchronized {
+  return _synchronized;
+}
+
+- (void)onDidFinishSynchronizing:(void (^)())completion {
+  ASDisplayNodeAssertMainThread();
+  if ([self isSynchronized]) {
+    ASPerformBlockOnMainThread(completion);
+  } else {
+    // Hang on to the completion block so that it gets called the next time view is synchronized to data.
+    [_onDidFinishSynchronizingBlocks addObject:[completion copy]];
+  }
+}
+
 - (void)updateWithChangeSet:(_ASHierarchyChangeSet *)changeSet
 {
   ASDisplayNodeAssertMainThread();
+
+  _synchronized = NO;
+
+  [changeSet addCompletionHandler:^(BOOL finished) {
+    self->_synchronized = YES;
+    [self onDidFinishProcessingUpdates:^{
+      if (self->_synchronized) {
+        for (ASDataControllerSynchronizationBlock block in self->_onDidFinishSynchronizingBlocks) {
+          block();
+        }
+        [_onDidFinishSynchronizingBlocks removeAllObjects];
+      }
+    }];
+  }];
   
   if (changeSet.includesReloadData) {
     if (_initialReloadDataHasBeenCalled) {
@@ -558,6 +591,7 @@ typedef dispatch_block_t ASDataControllerCompletionBlock;
   as_log_debug(ASCollectionLog(), "New content: %@", newMap.smallDescription);
 
   Class<ASDataControllerLayoutDelegate> layoutDelegateClass = [self.layoutDelegate class];
+  ++_editingTransactionQueueCount;
   dispatch_group_async(_editingTransactionGroup, _editingTransactionQueue, ^{
     __block __unused os_activity_scope_state_s preparationScope = {}; // unused if deployment target < iOS10
     as_activity_scope_enter(as_activity_create("Prepare nodes for collection update", AS_ACTIVITY_CURRENT, OS_ACTIVITY_FLAG_DEFAULT), &preparationScope);
@@ -577,6 +611,7 @@ typedef dispatch_block_t ASDataControllerCompletionBlock;
           self.visibleMap = newMap;
         }];
       }];
+      --self->_editingTransactionQueueCount;
     };
 
     // Step 3: Call the layout delegate if possible. Otherwise, allocate and layout all elements
@@ -584,9 +619,17 @@ typedef dispatch_block_t ASDataControllerCompletionBlock;
       [layoutDelegateClass calculateLayoutWithContext:layoutContext];
       completion();
     } else {
-      NSArray<ASCollectionElement *> *elementsToProcess = ASArrayByFlatMapping(newMap,
-                                                                               ASCollectionElement *element,
-                                                                               (element.nodeIfAllocated.calculatedLayout == nil ? element : nil));
+      NSMutableArray<ASCollectionElement *> *elementsToProcess = [NSMutableArray array];
+      for (ASCollectionElement *element in newMap) {
+        ASCellNode *nodeIfAllocated = element.nodeIfAllocated;
+        if (nodeIfAllocated.shouldUseUIKitCell) {
+          // If the node exists and we know it is a passthrough cell, we know it will never have a .calculatedLayout.
+          continue;
+        } else if (nodeIfAllocated.calculatedLayout == nil) {
+          // If the node hasn't been allocated, or it doesn't have a valid layout, let's process it.
+          [elementsToProcess addObject:element];
+        }
+      }
       [self _allocateNodesFromElements:elementsToProcess completion:completion];
     }
   });
@@ -820,6 +863,13 @@ typedef dispatch_block_t ASDataControllerCompletionBlock;
       }
     }];
   });
+}
+
+- (void)clearData
+{
+  ASDisplayNodeAssertMainThread();
+  [self waitUntilAllUpdatesAreProcessed];
+  self.visibleMap = self.pendingMap = [[ASElementMap alloc] init];
 }
 
 # pragma mark - Helper methods